# Bump to latest SMD to pick up performance optimizations