… go.sum
… staging/src/k8s.io/api/go.sum
… staging/src/k8s.io/apiextensions-apiserver/go.sum
… staging/src/k8s.io/apimachinery/go.sum
… staging/src/k8s.io/apiserver/go.sum
… staging/src/k8s.io/cli-runtime/go.sum
… staging/src/k8s.io/client-go/go.sum
… staging/src/k8s.io/cloud-provider/go.sum
… staging/src/k8s.io/cluster-bootstrap/go.sum
… staging/src/k8s.io/component-base/go.sum
… staging/src/k8s.io/csi-translation-lib/go.sum
… staging/src/k8s.io/kube-aggregator/go.sum
… staging/src/k8s.io/kube-controller-manager/go.sum
… staging/src/k8s.io/kube-proxy/go.sum
… staging/src/k8s.io/kube-scheduler/go.sum
… staging/src/k8s.io/kubectl/go.sum
… staging/src/k8s.io/kubelet/go.sum
… staging/src/k8s.io/legacy-cloud-providers/go.sum
… staging/src/k8s.io/metrics/go.sum
… staging/src/k8s.io/sample-apiserver/go.sum
… staging/src/k8s.io/sample-cli-plugin/go.sum
… staging/src/k8s.io/sample-controller/go.sum
… vendor/modules.txt
… vendor/sigs.k8s.io/structured-merge-diff/v3/fieldpath/fromvalue.go
… vendor/sigs.k8s.io/structured-merge-diff/v3/fieldpath/serialize.go
… vendor/sigs.k8s.io/structured-merge-diff/v3/typed/helpers.go
… vendor/sigs.k8s.io/structured-merge-diff/v3/typed/merge.go
… vendor/sigs.k8s.io/structured-merge-diff/v3/typed/remove.go
… vendor/sigs.k8s.io/structured-merge-diff/v3/typed/tofieldset.go
… vendor/sigs.k8s.io/structured-merge-diff/v3/typed/typed.go