**What type of PR is this?**

/kind cleanup

**What this PR does / why we need it**:

Decreases allocations and CPU usage of the code introduced by enabling server side apply.

benchstat of `FieldManager.Update` improvement using `fieldmanager.BenchmarkNewObject`:
```
name                           old time/op    new time/op    delta
NewObject/Pod/Update           587µs ± 1%     380µs ± 2%  -35.30%  (p=0.008 n=5+5)
NewObject/Node/Update          842µs ± 1%     540µs ± 1%  -35.85%  (p=0.008 n=5+5)
NewObject/Endpoints/Update    2.02ms ± 1%    2.14ms ± 2%   +6.15%  (p=0.008 n=5+5)

name                           old alloc/op   new alloc/op   delta
NewObject/Pod/Update           167kB ± 0%      79kB ± 0%  -52.94%  (p=0.008 n=5+5)
NewObject/Node/Update          231kB ± 0%     121kB ± 0%  -47.49%  (p=0.008 n=5+5)
NewObject/Endpoints/Update     198kB ± 0%     175kB ± 0%  -11.45%  (p=0.008 n=5+5)

name                           old allocs/op  new allocs/op  delta
NewObject/Pod/Update           2.83k ± 0%     1.71k ± 0%  -39.62%  (p=0.000 n=5+4)
NewObject/Node/Update          3.86k ± 0%     2.47k ± 0%  -36.02%  (p=0.008 n=5+5)
NewObject/Endpoints/Update     5.31k ± 0%     4.23k ± 0%  -20.46%  (p=0.008 n=5+5)
```

We also plan to run these tests with RunParallel enabled, which we expect to show a larger latency savings since some of the optimizations reduced lock contention that was adding latency to Updates.

**Which issue(s) this PR fixes**:

**Special notes for your reviewer**:

See the following PRs for details of the optimizations being enabled:
- https://github.com/kubernetes-sigs/structured-merge-diff/pull/136
- https://github.com/kubernetes-sigs/structured-merge-diff/pull/151
- https://github.com/kubernetes-sigs/structured-merge-diff/pull/152
- https://github.com/kubernetes-sigs/structured-merge-diff/pull/153
- https://github.com/kubernetes-sigs/structured-merge-diff/pull/155

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```

/sig api-machinery
cc @apelisse @jennybuckley @lavalamp @liggitt