@liggitt 2020-02-07
also a spurious cpu.prof file