@jpbetz 2020-02-07
Need to pick up https://github.com/kubernetes-sigs/structured-merge-diff/pull/156 once it merges. We missed a corner case in how Kubernetes marshals json.