@jpbetz 2020-02-07
Updated to pick up https://github.com/kubernetes-sigs/structured-merge-diff/pull/156

/hold cancel