@jpbetz 2020-02-07
/test pull-kubernetes-e2e-kind

Last failure (https://prow.k8s.io/view/gcs/kubernetes-jenkins/pr-logs/pull/87908/pull-kubernetes-e2e-kind/1225887937401458688):
```
[Testpattern: Pre-provisioned PV (default fs)]
should support readOnly directory specified in the volumeMount [It]
```