@liggitt 2020-02-08
looks like "exited with error code 1" failure started in kind ~1/7

https://storage.googleapis.com/k8s-gubernator/triage/index.html?date=2020-01-15&pr=1&text=command%20terminated%20with%20exit%20code%201%22